Refactor examples and add register payload. :rocket: :fire: :gem: :test_tube:
